% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{slab}
\alias{slab}
\title{Prepare for Averaging by Decade}
\usage{
slab(array, scheme, fun = "weighted mean")
}
\arguments{
\item{array}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}

\item{scheme}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}

\item{fun}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}
}
\value{
A dataframe containing a summarised month of NEMO-MEDUSA output, gaining a decade column, and dropping columns
which aren't needed for spatial maps.
}
\description{
This function cleans the saved NEMO-MEDUSA monthly summaries, for averaging into decades.
}
\concept{NEMO-MEDUSA averages}
