% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slabR-package.R
\docType{package}
\name{slabR-package}
\alias{slabR}
\alias{slabR-package}
\title{slabR: Summarise Slabs from Arrays in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create summary slabs from arrays according to a summary scheme. This is useful for ecosystem models which use a single value to represent volumes within a 3D grid. The different summaries are implemented in C++ using Rcpp for speed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-h-laverick.github.io/slabR/}
  \item \url{https:github.com/Jack-H-Laverick/slabR}
  \item Report bugs at \url{https:github.com/Jack-H-Laverick/slabR/issues}
}

}
\author{
\strong{Maintainer}: Jack H Laverick \email{jacklaverick@ymail.com} (\href{https://orcid.org/0000-0001-8829-2084}{ORCID})

}
\keyword{internal}
